# samples/S7.Meridian/docs/meridian-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Koan API",
    "version": "v1"
  },
  "paths": {
...
    "/api/pipelines/{pipelineId}/notes": {
      "get": {
        "tags": [
          "PipelineNotes"
        ],
        "parameters": [
          {
            "name": "pipelineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Koan-Trace-Id": {
                "description": "Trace correlation id for this response.",
                "schema": {
                  "type": "string"
                }
              },
              "Koan-InMemory-Paging": {
                "description": "true when in-memory pagination fallback occurred.",
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NotesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "Koan-Trace-Id": {
                "description": "Trace correlation id for this response.",
                "schema": {
                  "type": "string"
                }
              },
              "Koan-InMemory-Paging": {
                "description": "true when in-memory pagination fallback occurred.",
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PipelineNotes"
        ],
        "parameters": [
          {
            "name": "pipelineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotesRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotesRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotesRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Koan-Trace-Id": {
                "description": "Trace correlation id for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateNotesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateNotesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateNotesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {
              "Koan-Trace-Id": {
                "description": "Trace correlation id for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "Koan-Trace-Id": {
                "description": "Trace correlation id for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PipelineNotes"
        ],
        "parameters": [
          {
            "name": "pipelineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reProcess",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {
              "Koan-Trace-Id": {
                "description": "Trace correlation id for this response.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "Koan-Trace-Id": {
                "description": "Trace correlation id for this response.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NotesResponse": {
        "type": "object",
        "properties": {
          "pipelineId": {
            "type": "string",
            "nullable": true
          },
          "authoritativeNotes": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UpdateNotesRequest": {
        "type": "object",
        "properties": {
          "authoritativeNotes": {
            "type": "string",
            "nullable": true
          },
          "reProcess": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UpdateNotesResponse": {
        "type": "object",
        "properties": {
          "pipelineId": {
            "type": "string",
            "nullable": true
          },
          "authoritativeNotes": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reProcessScheduled": {
            "type": "boolean"
          },
          "jobId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}